Extract shared system-info microcrate for CLI/OpenCL
