+++ crates/bitnet-opencl/Cargo.toml
@@ -14,6 +14,7 @@ rust-version.workspace = true
 [dependencies]
 bitnet-common = { path = "../bitnet-common", version = "0.2.1-dev" }
 bitnet-spirv = { path = "../bitnet-spirv", version = "0.2.1-dev" }
+bitnet-system-info-core = { workspace = true }
 log.workspace = true
 thiserror.workspace = true
 serde = { workspace = true }
